Major Points Update
Completion of phase 1 of the points system.

Players now rank up depending on points earned.

Ranks are 1 to 100 and displayed as a chat prefix.

If possible serverside then I will look at getting the rank prefix displayed on the scoreboard also.

Counting the points earned marks a critical turning point in the points system as it allows for awards such as airstrikes and care packages to be earned from gameplay.

This update includes the points couinting and a new console command to view the current points awards by typing /points